+++ src/app/cart/models/cart-item.ts
@@ -3,4 +3,14 @@ export class CartItem {
   count = 0;
   total = 0;
   pricePerOne = 0;
+
+  decrease() {
+    this.count--;
+    this.total = this.pricePerOne * this.count;
+  }
+
+  increase() {
+    this.count++;
+    this.total = this.pricePerOne * this.count;
+  }
 }