+++ src/app/core/models/options.module.ts
@@ -1,0 +1,5 @@
+export class Options {
+  id: number;
+  login: string;
+  email: string;
+}