+++ src/app/shared/directives/click-highlight.directive.spec.ts
@@ -1,0 +1,8 @@
+import { ClickHighlightDirective } from './click-highlight.directive';
+
+describe('ClickHighlightDirective', () => {
+  it('should create an instance', () => {
+    const directive = new ClickHighlightDirective();
+    expect(directive).toBeTruthy();
+  });
+});