+++ src/app/shared/shared.module.ts
@@ -1,12 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HighlightDirective } from './directives/highlight.directive';
+import { ClickHighlightDirective } from './directives/click-highlight.directive';
 
 @NgModule({
-  declarations: [HighlightDirective],
+  declarations: [HighlightDirective, ClickHighlightDirective],
   imports: [
     CommonModule
   ],
-  exports: [HighlightDirective]
+  exports: [HighlightDirective, ClickHighlightDirective]
 })
 export class SharedModule { }