+++ changelog.md
@@ -1,3 +1,10 @@
+#Version 1.2
+- Added About component;
+- Added service for working with local storage;
+- Added directive which changing element border color on click;
+- Added constant service for getting application constants;
+- Added generator service for generating random string;
+
 # Version 1.1
 - Added "add product" feature;
 - Added cart item highlighting on hover;